return index instead of value

+++ src/Rmaxdd.c
@@ -7,6 +7,6 @@
 
 #include "maxdd.h"
 
-void bestbeforemaxdd(double *x, int *r, double *threshold, int *dir, double *results){
+void bestbeforemaxdd(double *x, int *r, double *threshold, int *dir, int *results){
     best_before_maxdd(x, r[0], threshold[0], dir[0] > 0 ? 1 : -1, results);
 }
